organize mir pretty.rs and move more things into it; move statement-related things out of mir/mod.rs
